python -m debug will start buche
